Prevent horrible scrolling/resizing on iOS
Get around the behavior on iOS where vh units refer to the "full" height
of a page which is only the height after scrolling has occurred. Now the
whole scroll-resizing is gone on iOS because the site is fixed to the
shorter page height. Not ideal but might be the best we can do?

All vh units are now replaced with percentages