fix buttonMessage when multiple configurations of a product are in cart
Before, productInCart and friends would always refer to the *first* 
configuration of a product in the cart. Now, it refers to the most 
recently added, so re-opening product-options for the same product will 
configure options to match what they were set to most recently. 

Additionally, the button text has been decoupled from these computed 
props, so it will properly display "added" if the selected configuration 
is anywhere in the cart